fastapi_app/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langgraph
langchain-openai
langchain
